Moved error validation to the community code base
