+++ packages/rocketchat-i18n/i18n/en.i18n.json
@@ -675,6 +675,7 @@
   "Clear_filters": "Clear filters",
   "clear_history": "Clear History",
   "Click_here": "Click here",
+  "Click_here_for_more_details_or_contact_sales_for_a_new_license": "<a target=\"_blank\" href=\"__url__\">Click here</a> for more details or contact <strong>__email__</strong> for a new license.",
   "Click_here_for_more_info": "Click here for more info",
   "Click_here_to_enter_your_encryption_password": "Click here to enter your encryption password",
   "Click_here_to_view_and_copy_your_password": "Click here to view and copy your password.",
@@ -2279,6 +2280,7 @@
   "Max_number_incoming_livechats_displayed_description": "(Optional) Max number of items displayed in the incoming Omnichannel queue.",
   "Max_number_of_uses": "Max number of uses",
   "Maximum": "Maximum",
+  "Maximum_number_of_guests_reached": "Maximum number of guests reached",
   "Media": "Media",
   "Medium": "Medium",
   "Members": "Members",
@@ -3738,6 +3740,7 @@
   "You_need_to_type_in_your_username_in_order_to_do_this": "You need to type in your username in order to do this!",
   "You_need_to_verifiy_your_email_address_to_get_notications": "You need to verify your email address to get notifications",
   "You_need_to_write_something": "You need to write something!",
+  "You_reached_the_maximum_number_of_guest_users_allowed_by_your_license": "You reached the maximum number of guest users allowed by your license.",
   "You_should_inform_one_url_at_least": "You should define at least one URL.",
   "You_should_name_it_to_easily_manage_your_integrations": "You should name it to easily manage your integrations.",
   "You_will_not_be_able_to_recover": "You will not be able to recover this message!",
